Update POSsystem.cpp
Percentage function parameter data type from int to float

+++ POSsystem.cpp
@@ -6,15 +6,15 @@ using namespace std;
 
 namespace calculate {
 
-    float calculateTip(float gross, int percentage){
+    float calculateTip(float gross, float percentage){
         return gross * (percentage/100);
     }
 
-    float calculateNet(float tip, float gross, int percentage) {
+    float calculateNet(float tip, float gross, float percentage) {
         return gross + calculateTip(gross, percentage);
     }
 
-    void displayReciept(float gross, float net, float tip, int percentage) {
+    void displayReciept(float gross, float net, float tip, float percentage) {
         std::cout << "Your Gross Amount is: " << gross << endl;
         std::cout << "=============================================" << endl;
         std::cout << "Percentage for tip: " << percentage << endl;
